refactor: delete unused components
